chore: disable CORS on development setup
We need to disable CORS to allow developer to play with the API with tools such as Apollo Studio.
We could also add more domains to the CORS but we favor the most simple solution in this case.
Final security compliance once deployed in kubernetes cluster is not impacted by this change.
The development boolean is only defined by Webpack when running development builds.